Force floats for volume image plotting

+++ src/types/VolumeImage/Plotting.jl
@@ -2,9 +2,9 @@ function plot(vi::VolumeImage; ncols::Int = 2, colorbar::Bool = true, colorbar_t
     max_size::Union(Number, Nothing) = 1, min_size = 0.2,
     threshold::Number = unique(sort(vec(squeeze(mean(vi.data, 4), 4))))[2], threshold_ratio::Number = 1/1000, kwargs...)
 
-    x = [float(xi) * 1000 for xi in vi.x]
-    y = [float(yi) * 1000 for yi in vi.y]
-    z = [float(zi) * 1000 for zi in vi.z]
+    x = float([float(xi) * 1000 for xi in vi.x])
+    y = float([float(yi) * 1000 for yi in vi.y])
+    z = float([float(zi) * 1000 for zi in vi.z])
     s = squeeze(mean(vi.data, 4), 4)
 
     plot_dat(x, y, z, s, threshold_ratio = threshold_ratio, ncols = ncols,
